third_party/blink/web_tests/platform/mac-mac-arm11.0/external/wpt/css/css-fonts/generic-family-keywords-002-expected.txt

This is a testharness.js-based test.
FAIL font-family: -webkit-serif treated as <font-family>, not <generic-name> assert_equals: expected 50 but got 25
FAIL font-family: -webkit-sans-serif treated as <font-family>, not <generic-name> assert_equals: expected 50 but got 28
FAIL font-family: -webkit-cursive treated as <font-family>, not <generic-name> assert_equals: expected 50 but got 26
FAIL font-family: -webkit-fantasy treated as <font-family>, not <generic-name> assert_equals: expected 50 but got 33
FAIL font-family: -webkit-monospace treated as <font-family>, not <generic-name> assert_equals: expected 50 but got 30
PASS font-family: -webkit-system treated as <font-family>, not <generic-name>
PASS font-family: -webkit-emoji treated as <font-family>, not <generic-name>
PASS font-family: -webkit-math treated as <font-family>, not <generic-name>
PASS font-family: -webkit-fangsong treated as <font-family>, not <generic-name>
PASS font-family: -webkit-ui-serif treated as <font-family>, not <generic-name>
PASS font-family: -webkit-ui-sans-serif treated as <font-family>, not <generic-name>
PASS font-family: -webkit-ui-monospace treated as <font-family>, not <generic-name>
PASS font-family: -webkit-ui-rounded treated as <font-family>, not <generic-name>
FAIL font-family: -webkit-body treated as <font-family>, not <generic-name> assert_equals: expected 50 but got 25
FAIL font-family: -webkit-standard treated as <font-family>, not <generic-name> assert_equals: expected 50 but got 25
FAIL font-family: -webkit-pictograph treated as <font-family>, not <generic-name> assert_equals: expected 50 but got 25
Harness: the test ran to completion.
